backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.9.0
pydantic-settings==2.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
boto3
requests
slowapi==0.1.9
redis
structlog
